Разбил код на модули, создал точку входа main
